Better BC. Fix issues with identification of current cart. Misc.
- When a filter is enabling the shipping purchase requirement, and the cart
is current, force ITE_Cart::requires_shipping() to true.
- Don't destory session model in ITE_Line_Item_Repository::destory() if
purchased_at flag is set. Set purchased_at flag if not set when adding a
transaction and gateway requires the cart after purchase.
- Fix test pollution.

+++ lib/cart/line-items/class.cached-session-repository.php
@@ -227,7 +227,7 @@ class ITE_Line_Item_Cached_Session_Repository extends ITE_Line_Item_Session_Repo
 	public function destroy( $cart_id ) {
 		$model = ITE_Session_Model::from_cart_id( $cart_id );
 
-		if ( $model ) {
+		if ( $model && ! $model->purchased_at ) {
 			$model->delete();
 		}
